refactor: 파일 명 및 파일 구조 변경
