+++ script/controller/controller.js
@@ -1,4 +1,4 @@
-import { insertNewsData } from "../view/rollingBar.js";
+import { insertNewsData } from "../view/newsRollingBar.js";
 const rollingController = () => {
   fetch("http://localhost:3001/rollingData")
     .then((res) => res.json())
@@ -8,11 +8,11 @@ const rollingController = () => {
     })
     .catch((error) => console.error("Error:", error));
 };
-const deliverData = (data, num1, num2, className) => {
+const deliverData = (data, startData, endData, rollingBox) => {
   const newsData = [];
-  for (let i = num1; i < num2; i++) {
+  for (let i = startData; i < endData; i++) {
     newsData.push(data[i].headLine);
   }
-  insertNewsData(newsData, className);
+  insertNewsData(newsData, rollingBox);
 };
 export { rollingController };
