+++ script/view/newsRollingBar.js
